add buffer.rs and hex.rs
